fix: parse prices as floats

+++ src/CartItem.jsx
@@ -10,7 +10,7 @@ const CartItem = ({ onContinueShopping }) => {
   // Calculate total amount for all products in the cart
   const calculateTotalAmount = () =>
     cart.reduce(
-      (acc, item) => acc + item.quantity * parseInt(item.cost.substring(1)),
+      (acc, item) => acc + item.quantity * parseFloat(item.cost.substring(1)),
       0,
     );
 
@@ -32,7 +32,7 @@ const CartItem = ({ onContinueShopping }) => {
 
   // Calculate total cost based on quantity for an item
   const calculateTotalCost = (item) =>
-    item.quantity * parseInt(item.cost.substring(1));
+    item.quantity * parseFloat(item.cost.substring(1));
 
   return (
     <div className="cart-container">
